implement timedeltas.test_rendering
