+++ pandas/tests/scalar/test_timedelta.py
@@ -270,17 +270,6 @@ class TestTimedeltas(object):
         rng = Timedelta(np.nan)
         assert np.isnan(rng.total_seconds())
 
-    def test_repr(self):
-
-        assert (repr(Timedelta(10, unit='d')) ==
-                "Timedelta('10 days 00:00:00')")
-        assert (repr(Timedelta(10, unit='s')) ==
-                "Timedelta('0 days 00:00:10')")
-        assert (repr(Timedelta(10, unit='ms')) ==
-                "Timedelta('0 days 00:00:00.010000')")
-        assert (repr(Timedelta(-10, unit='ms')) ==
-                "Timedelta('-1 days +23:59:59.990000')")
-
     def test_conversion(self):
 
         for td in [Timedelta(10, unit='d'),
@@ -735,37 +724,3 @@ class TestTimedeltas(object):
         result = s.dt.components
         assert not result.iloc[0].isna().all()
         assert result.iloc[1].isna().all()
-
-    def test_isoformat(self):
-        td = Timedelta(days=6, minutes=50, seconds=3,
-                       milliseconds=10, microseconds=10, nanoseconds=12)
-        expected = 'P6DT0H50M3.010010012S'
-        result = td.isoformat()
-        assert result == expected
-
-        td = Timedelta(days=4, hours=12, minutes=30, seconds=5)
-        result = td.isoformat()
-        expected = 'P4DT12H30M5S'
-        assert result == expected
-
-        td = Timedelta(nanoseconds=123)
-        result = td.isoformat()
-        expected = 'P0DT0H0M0.000000123S'
-        assert result == expected
-
-        # trim nano
-        td = Timedelta(microseconds=10)
-        result = td.isoformat()
-        expected = 'P0DT0H0M0.00001S'
-        assert result == expected
-
-        # trim micro
-        td = Timedelta(milliseconds=1)
-        result = td.isoformat()
-        expected = 'P0DT0H0M0.001S'
-        assert result == expected
-
-        # don't strip every 0
-        result = Timedelta(minutes=1).isoformat()
-        expected = 'P0DT0H1M0S'
-        assert result == expected
